requirements.txt
python-dotenv~=1.1.0
kafka-python~=2.1.4
anyio~=4.9.0